backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dj-database-url==2.1.0
gunicorn==21.2.0
pillow==12.0.0
psycopg[binary]==3.2.3
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
whitenoise==6.6.0
